Added func<float, float> for bezier, Tweening2.
Tweening2 works off of storing the object and using a property info. It's sleeker to use (no more setter delegates) but for objects with no LerpFunction it's more difficult (no more using internal float values through setter). Either going to choose, or provice both as options for when you need one or the other.